fix(audit): use uint256 for unbounded batch loops
uint8 counters panic at length >255 (M-26). Hygiene only — no
attacker profit — but removes implicit batch footgun.

+++ contracts/facets/PartyA/PartyAFacetImpl.sol
@@ -102,7 +102,7 @@ library PartyAFacetImpl {
 		require(MpcCore.decrypt(balanceSufficient), "PartyAFacet: insufficient available balance");
 
 		// Additional non-encrypted validations
-		for (uint8 i = 0; i < partyBsWhiteList.length; i++) {
+		for (uint256 i = 0; i < partyBsWhiteList.length; i++) {
 			require(partyBsWhiteList[i] != msg.sender, "PartyAFacet: Sender isn't allowed in partyBWhiteList");
 		}
 		require(maLayout.affiliateStatus[affiliate] || affiliate == address(0), "PartyAFacet: Invalid affiliate");
